Relocate function for filtering definitions
